Extract todoReducer.ts
it has reducer function and actions
